malcheck-dash/requirements.txt
click==8.1.3
Faker==17.6.0
Flask==2.2.3
Flask-HTTPAuth==4.7.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
minio==7.1.13
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.5.post1
text-unidecode==1.3
Werkzeug==2.2.3